Frame counter must be int32
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@12644 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libsphinxbase/fe/fe_internal.h
@@ -155,7 +155,7 @@ struct fe_s {
     float32 pre_emphasis_alpha;
     int32 seed;
 
-    int16 frame_counter;
+    int32 frame_counter;
     uint8 start_flag;
     uint8 reserved;
